Bug 2024075 Test Version Refactor: Moved subtest columns to test version strategy and test version files (#1017)
moved subtest columns to test version strategy and test version files

Updated tests to make mann-whitney-u the default. Added duplicated code to revisionRowHelpers.
