+++ src/utils/revisionRowHelpers.ts
@@ -1,3 +1,9 @@
+export function determineSign(baseValue: number, newValue: number) {
+  if (baseValue > newValue) return '>';
+  if (baseValue < newValue) return '<';
+  return '';
+}
+
 export function determineStatus(improvement: boolean, regression: boolean) {
   if (improvement) return 'Improvement';
   if (regression) return 'Regression';
